
\name{print.extmod}
\alias{print.extmod}
\alias{print.simpextmod}
\title{
Result style
}
\description{
Defines how extinction predictor results are displayed.
}
\usage{
\method{print}{extmod}(x,\dots)
\method{print}{simpextmod}(x,\dots)
}

\arguments{

  \item{x}{results from the estimator}
  \item{\dots}{...}
}
